fix tip type

+++ goodfood-orders/goodfood-orders/Models/CreateOrderModel.cs
@@ -4,6 +4,6 @@
     {
         public int RestaurantId { get; set; }
         public int UserId { get; set; }
-        public int Tip { get; set; }
+        public double Tip { get; set; }
     }
 }
